+++ espnow_test.h
@@ -1,0 +1,10 @@
+#ifndef ESPNOW_TEST_H
+#define ESPNOW_TEST_H
+
+namespace EspNowTest {
+  void espNowTestSetup(bool broadcastMode);
+  void espNowTestLoop();
+  void espNowTestCleanup();
+}
+
+#endif  // ESPNOW_TEST_H